+++ packages/firestore/test/unit/local/counting_query_engine.ts
@@ -17,6 +17,7 @@
 
 import { Query } from '../../../src/core/query';
 import { SnapshotVersion } from '../../../src/core/snapshot_version';
+import { DocumentOverlayCache } from '../../../src/local/document_overlay_cache';
 import { IndexManager } from '../../../src/local/index_manager';
 import { LocalDocumentsView } from '../../../src/local/local_documents_view';
 import { MutationQueue } from '../../../src/local/mutation_queue';
@@ -24,7 +25,14 @@ import { PersistencePromise } from '../../../src/local/persistence_promise';
 import { PersistenceTransaction } from '../../../src/local/persistence_transaction';
 import { QueryEngine } from '../../../src/local/query_engine';
 import { RemoteDocumentCache } from '../../../src/local/remote_document_cache';
-import { DocumentKeySet, DocumentMap } from '../../../src/model/collections';
+import {
+  DocumentKeySet,
+  DocumentMap,
+  OverlayMap
+} from '../../../src/model/collections';
+import { DocumentKey } from '../../../src/model/document_key';
+import { Overlay } from '../../../src/model/overlay';
+import { ResourcePath } from '../../../src/model/path';
 
 /**
  * A test-only query engine that forwards all API calls and exposes the number
@@ -58,11 +66,25 @@ export class CountingQueryEngine extends QueryEngine {
    */
   documentsReadByKey = 0;
 
+  /**
+   * The number of documents returned by the OverlayCache's `getOverlays()`
+   * API (since the last call to `resetCounts()`)
+   */
+  overlaysReadByCollection = 0;
+
+  /**
+   * The number of documents returned by the OverlayCache's `getOverlay()`
+   * APIs (since the last call to `resetCounts()`)
+   */
+  overlaysReadByKey = 0;
+
   resetCounts(): void {
     this.mutationsReadByCollection = 0;
     this.mutationsReadByKey = 0;
     this.documentsReadByCollection = 0;
     this.documentsReadByKey = 0;
+    this.overlaysReadByCollection = 0;
+    this.overlaysReadByKey = 0;
   }
 
   getDocumentsMatchingQuery(
@@ -86,7 +108,7 @@ export class CountingQueryEngine extends QueryEngine {
     const view = new LocalDocumentsView(
       this.wrapRemoteDocumentCache(localDocuments.remoteDocumentCache),
       this.wrapMutationQueue(localDocuments.mutationQueue),
-      localDocuments.documentOverlayCache,
+      this.wrapDocumentOverlayCache(localDocuments.documentOverlayCache),
       localDocuments.indexManager
     );
     return super.initialize(view, indexManager);
@@ -127,13 +149,17 @@ export class CountingQueryEngine extends QueryEngine {
       },
       getEntries: (transaction, documentKeys) => {
         return subject.getEntries(transaction, documentKeys).next(result => {
-          this.documentsReadByKey += result.size;
+          result.forEach((key, doc) => {
+            if (doc.isValidDocument()) {
+              this.documentsReadByKey++;
+            }
+          });
           return result;
         });
       },
       getEntry: (transaction, documentKey) => {
         return subject.getEntry(transaction, documentKey).next(result => {
-          this.documentsReadByKey += result ? 1 : 0;
+          this.documentsReadByKey += result?.isValidDocument() ? 1 : 0;
           return result;
         });
       },
@@ -185,6 +211,59 @@ export class CountingQueryEngine extends QueryEngine {
       lookupMutationBatch: subject.lookupMutationBatch,
       performConsistencyCheck: subject.performConsistencyCheck,
       removeMutationBatch: subject.removeMutationBatch
+    };
+  }
+
+  private wrapDocumentOverlayCache(
+    subject: DocumentOverlayCache
+  ): DocumentOverlayCache {
+    return {
+      getOverlay: (
+        transaction: PersistenceTransaction,
+        key: DocumentKey
+      ): PersistencePromise<Overlay | null> => {
+        this.overlaysReadByKey++;
+        return subject.getOverlay(transaction, key);
+      },
+      getOverlays: (
+        transaction: PersistenceTransaction,
+        keys: DocumentKey[]
+      ): PersistencePromise<OverlayMap> => {
+        this.overlaysReadByKey += keys.length;
+        return subject.getOverlays(transaction, keys);
+      },
+      getOverlaysForCollection: (
+        transaction: PersistenceTransaction,
+        collection: ResourcePath,
+        sinceBatchId: number
+      ): PersistencePromise<OverlayMap> => {
+        return subject
+          .getOverlaysForCollection(transaction, collection, sinceBatchId)
+          .next(result => {
+            this.overlaysReadByCollection += result.size();
+            return result;
+          });
+      },
+      getOverlaysForCollectionGroup: (
+        transaction: PersistenceTransaction,
+        collectionGroup: string,
+        sinceBatchId: number,
+        count: number
+      ): PersistencePromise<OverlayMap> => {
+        return subject
+          .getOverlaysForCollectionGroup(
+            transaction,
+            collectionGroup,
+            sinceBatchId,
+            count
+          )
+          .next(result => {
+            this.overlaysReadByCollection += result.size();
+            return result;
+          });
+      },
+      removeOverlaysForBatchId: subject.removeOverlaysForBatchId,
+      saveOverlays: subject.saveOverlays
     };
   }
 }